Update: New Mobile Sidebar.
Using the slide in modal for the mobile sidebar.
