+++ components/Utility/Modal/slidein.tsx
@@ -1,0 +1,103 @@
+import { Dispatch, Fragment, SetStateAction } from "react";
+import { Dialog, Transition } from "@headlessui/react";
+
+type Props = {
+  open: boolean;
+  setOpen: Dispatch<SetStateAction<boolean>>;
+  title: string;
+  theme?: string;
+  children: JSX.Element;
+};
+
+export default function SlideIn({
+  open,
+  setOpen,
+  title,
+  theme,
+  children,
+}: Props) {
+  return (
+    <Transition.Root show={open} as={Fragment}>
+      <Dialog as="div" className="relative z-10" onClose={setOpen}>
+        <Transition.Child
+          as={Fragment}
+          enter="ease-in-out duration-500"
+          enterFrom="opacity-0"
+          enterTo="opacity-100"
+          leave="ease-in-out duration-500"
+          leaveFrom="opacity-100"
+          leaveTo="opacity-0">
+          <div className="fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity" />
+        </Transition.Child>
+
+        <div className="fixed inset-0 overflow-hidden">
+          <div className="absolute inset-0 overflow-hidden">
+            <div className="pointer-events-none fixed inset-y-0 right-0 flex max-w-full pl-10">
+              <Transition.Child
+                as={Fragment}
+                enter="transform transition ease-in-out duration-500 sm:duration-700"
+                enterFrom="translate-x-full"
+                enterTo="translate-x-0"
+                leave="transform transition ease-in-out duration-500 sm:duration-700"
+                leaveFrom="translate-x-0"
+                leaveTo="translate-x-full">
+                <Dialog.Panel
+                  className={`pointer-events-auto relative w-screen max-w-md`}>
+                  <Transition.Child
+                    as={Fragment}
+                    enter="ease-in-out duration-500"
+                    enterFrom="opacity-0"
+                    enterTo="opacity-100"
+                    leave="ease-in-out duration-500"
+                    leaveFrom="opacity-100"
+                    leaveTo="opacity-0">
+                    <div className="absolute top-24 left-0 -ml-8 flex pt-4 pr-2 sm:-ml-10 sm:pr-4">
+                      <button
+                        type="button"
+                        className="rounded-md text-gray-300 hover:text-white focus:outline-none focus:ring-2 focus:ring-white"
+                        onClick={() => setOpen(false)}>
+                        <span className="sr-only">Close panel</span>
+                        <svg
+                          xmlns="http://www.w3.org/2000/svg"
+                          fill="none"
+                          viewBox="0 0 24 24"
+                          strokeWidth={1.5}
+                          stroke="currentColor"
+                          className="w-6 h-6">
+                          <path
+                            strokeLinecap="round"
+                            strokeLinejoin="round"
+                            d="M6 18L18 6M6 6l12 12"
+                          />
+                        </svg>
+                      </button>
+                    </div>
+                  </Transition.Child>
+                  <div
+                    className={`flex h-full mt-20 flex-col py-6 shadow-xl
+                   ${
+                     theme === "light"
+                       ? "bg-task-light-white"
+                       : "bg-task-sidebar-dark"
+                   }
+                  `}>
+                    <div className="px-4 sm:px-6">
+                      <Dialog.Title className="text-lg font-medium text-gray-900">
+                        {title}
+                      </Dialog.Title>
+                    </div>
+                    <div className="relative flex-1 px-4 sm:px-6 overflow-y-scroll">
+                      <div className="absolute inset-0 px-4 sm:px-6">
+                        {children}
+                      </div>
+                    </div>
+                  </div>
+                </Dialog.Panel>
+              </Transition.Child>
+            </div>
+          </div>
+        </div>
+      </Dialog>
+    </Transition.Root>
+  );
+}
