+++ styles/globals.css
@@ -82,6 +82,9 @@
   padding: 0;
   margin: 0;
 }
+::-webkit-scrollbar {
+  display: none;
+}
 
 html,
 body {
@@ -149,7 +152,7 @@ body {
     top: 70px;
   }
 
-  .sidebar {
+  .hide-sidebar {
     display: none;
   }
